feat: replace kill command

+++ src/main/kotlin/model/command/KillCommand.kt
@@ -3,7 +3,7 @@ package model.command
 import model.command.factory.KillCommandFactory
 
 class KillCommand(private val factory: KillCommandFactory) {
-    fun run(pid: Int): Process {
+    fun run(pid: Long): Process {
         val command = factory.create(pid)
         return ProcessBuilder(command).start()
     }
